requirements.txt
-e git://github.com/smaiyaki/django.git@1.8.3mongoengine_patched#egg=django
Pillow==2.7.0
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.19
bleach==1.4
boto==2.38.0
celery==3.1.18
demjson==2.2.2
-e git://github.com/j1z0/django-rest-framework-mongoengine.git@drf_2_support#egg=rest_framework_mongo
djangorestframework==2.4.4
djangotoolbox==1.6.2
django-extensions==1.5.5
django-storages==1.1.8
geojson==1.0.9
html5lib==0.999
httplib2==0.9
kombu==3.0.24
mock==1.0.1
mongoengine==0.8.7.1
pygeocoder==1.2.5
pymongo==2.7.2
pyshp==1.2.1
pytz==2014.10
redis==2.10.3
requests==2.5.1
responses==0.3.0
six==1.8.0
wsgiref==0.1.2
django-nose==1.4
django-summernote==0.5.15
selenium==2.44.0
python-dateutil==2.4.2
